Update promise rules
With async/await most direct uses of promises will be carefully crafted.
Thus rules like promise/param-names and promise/always-return and
promise/catch-or-return tend to get in the way. Furthermore, unhandled
rejections are now logged and when developing applications it's good
practice to crash when they occur, again reducing the need for
promise/catch-or-return.

Add new rules promise/no-return-in-finally and promise/no-return-wrap
which are good practice even for those edge cases where promises are
crafted directly.